cobol_agentic_kg/requirements.txt
# Core Dependencies (flexible versions for uv compatibility)
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-neo4j>=0.2.0,<0.3.0
langchain-community>=0.3.0,<0.4.0
langchain-experimental>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0

# Database
neo4j>=5.0.0,<6.0.0

# Web Framework
streamlit>=1.40.0
streamlit-agraph>=0.0.45

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Git integration
GitPython>=3.1.0

# Progress tracking
tqdm>=4.60.0

# Async processing
aiofiles>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Monitoring
prometheus-client>=0.21.0

# File handling
chardet>=5.0.0

# Caching
redis>=5.0.0
hiredis>=2.3.0
